pyproject.toml
[tool.poetry]
name = "genapi"
version = "0.1.0"
description = "Powerful api for image and text generation."
authors = ["snehmehta <snehmehta933@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
django-unfold = "^0.8.0"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
werkzeug = "^2.3.7"
requests = "^2.31.0"
together = "^0.1.8"
discord-py-interactions = "^5.10.0"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
